Extract `ViewIntentFinder` from `AttachmentController`
